requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via edge-tts
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via typer
anyio==4.12.1
    # via httpx
attrs==25.4.0
    # via aiohttp
av==16.1.0
    # via faster-whisper
certifi==2026.1.4
    # via
    #   edge-tts
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via typer
colorama==0.4.6
    # via
    #   click
    #   tqdm
ctranslate2==4.7.1
    # via faster-whisper
edge-tts==7.2.7
    # via magic-paste (pyproject.toml)
faster-whisper==1.2.1
    # via magic-paste (pyproject.toml)
filelock==3.20.3
    # via huggingface-hub
flatbuffers==25.12.19
    # via onnxruntime
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2026.2.0
    # via huggingface-hub
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via huggingface-hub
huggingface-hub==1.4.1
    # via
    #   faster-whisper
    #   tokenizers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mouseinfo==0.1.3
    # via pyautogui
mpmath==1.3.0
    # via sympy
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
numpy==2.4.2
    # via
    #   magic-paste (pyproject.toml)
    #   ctranslate2
    #   onnxruntime
    #   scipy
onnxruntime==1.24.1
    # via faster-whisper
packaging==26.0
    # via
    #   huggingface-hub
    #   onnxruntime
pillow==12.1.0
    # via pyscreeze
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.5
    # via onnxruntime
pyaudio==0.2.14
    # via magic-paste (pyproject.toml)
pyautogui==0.9.54
    # via magic-paste (pyproject.toml)
pygame==2.6.1
    # via magic-paste (pyproject.toml)
pygetwindow==0.0.9
    # via pyautogui
pygments==2.19.2
    # via rich
pymsgbox==2.0.1
    # via pyautogui
pynput==1.8.1
    # via magic-paste (pyproject.toml)
pyperclip==1.11.0
    # via
    #   magic-paste (pyproject.toml)
    #   mouseinfo
pyrect==0.2.0
    # via pygetwindow
pyscreeze==1.0.1
    # via pyautogui
pyside6==6.10.2
    # via magic-paste (pyproject.toml)
pyside6-addons==6.10.2
    # via pyside6
pyside6-essentials==6.10.2
    # via
    #   pyside6
    #   pyside6-addons
pytweening==1.2.0
    # via pyautogui
pywin32==311
    # via magic-paste (pyproject.toml)
pyyaml==6.0.3
    # via
    #   ctranslate2
    #   huggingface-hub
requests==2.32.5
    # via magic-paste (pyproject.toml)
rich==14.3.2
    # via typer
scipy==1.17.0
    # via magic-paste (pyproject.toml)
setuptools==82.0.0
    # via ctranslate2
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
shiboken6==6.10.2
    # via
    #   pyside6
    #   pyside6-addons
    #   pyside6-essentials
six==1.17.0
    # via pynput
sympy==1.14.0
    # via onnxruntime
tabulate==0.9.0
    # via edge-tts
tokenizers==0.22.2
    # via faster-whisper
tqdm==4.67.3
    # via
    #   faster-whisper
    #   huggingface-hub
typer==0.23.0
    # via typer-slim
typer-slim==0.23.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   edge-tts
    #   huggingface-hub
urllib3==2.6.3
    # via requests
yarl==1.22.0
    # via aiohttp
